\name{SpatialFA-package}
\alias{SpatialFA-package}
\alias{SpatialFA}
\docType{package}
\title{
Package for implementing spatial factor analysis (SFA), a spatial analogue of conventional factor analysis
}
\description{
This package implements spatial factor analysis, which replaces factors with Gaussian random fields.  SFA provides a parsimonious (rank-reduced) model for spatial variation of multiple measures, e.g., species densities.
}
\details{
\tabular{ll}{
Package: \tab SpatialFA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-10-30\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
